+++ Tanks/src/entity/weapon/GrenadeGun.java
@@ -8,7 +8,7 @@ public class GrenadeGun implements Weapon {
 	private int cooldownTime = 30;
 	private int currentCooldown = 0;
 	
-	public void fire(double x, double y, BasicLevel level, double speedPercent, int angle) {
+	public void fire(double x, double y, BasicLevel level, double speedPercent, double angle) {
 		if(currentCooldown <= 0){
 		level.addEntity(new Grenade(x, y , level, speedPercent, angle));
 		currentCooldown = cooldownTime;
