+++ Tanks/src/entity/weapon/ShellGun.java
@@ -10,7 +10,7 @@ public class ShellGun implements Weapon {
 	int currentCooldown = 0;
 	
 	@Override
-	public void fire(double x, double y, BasicLevel level, double speedPercent, int angle) {
+	public void fire(double x, double y, BasicLevel level, double speedPercent, double angle) {
 		if(currentCooldown <= 0){
 		level.addEntity(new Shell(x, y , level, speedPercent, angle));
 		currentCooldown = cooldownTime;
