+++ Tanks/src/entity/weapon/Weapon.java
@@ -4,6 +4,6 @@ import entity.Projectile;
 import level.BasicLevel;
 
 public interface Weapon {
-	public void fire(double x, double y, BasicLevel level, double speedPercent, int angle);
+	public void fire(double x, double y, BasicLevel level, double speedPercent, double angle);
 	public void tick();
 }
